idsp::iir::coefficients

Struct Filter

pub struct Filter<T> {
    pub frequency: T,
    pub gain: T,
    pub shelf: T,
    pub shape: Shape<T>,
}

Standard audio biquad filter builder

https://www.w3.org/TR/audio-eq-cookbook/

Fields
frequency: T

Angular critical frequency (in units of sampling frequency), Corner frequency, or 3dB cutoff frequency. frequency=π is Nyquist.

gain: T

Passband gain

shelf: T

Shelf gain (only for peaking, lowshelf, highshelf) Relative to passband gain

shape: Shape<T>

Transition/corner shape

Implementations

impl<T> Filter<T>
where T: 'static + Float + FloatConst, f32: AsPrimitive<T>,

pub fn frequency( &mut self, critical_frequency: T, sample_frequency: T, ) -> &mut Self

Set crititcal frequency from absolute units.

Arguments

critical_frequency: “Relevant” or “corner” or “center” frequency in the same units as sample_frequency

sample_frequency: The sample frequency in the same units as critical_frequency. E.g. both in SI Hertz or rad/s.

pub fn critical_frequency(&mut self, f0: T) -> &mut Self

Set relative critical frequency

Arguments

f0: Relative critical frequency in units of the sample frequency. Must be 0 <= f0 <= 0.5.

pub fn angular_critical_frequency(&mut self, w0: T) -> &mut Self

Set relative critical angular frequency

Arguments

w0: Relative critical angular frequency. Must be 0 <= w0 <= π. Defaults to 0.0.

pub fn gain(&mut self, k: T) -> &mut Self

Set reference gain

Arguments

k: Linear reference gain. Defaults to 1.0.

pub fn gain_db(&mut self, k_db: T) -> &mut Self

Set reference gain in dB

Arguments

k_db: Reference gain in dB. Defaults to 0.0.

pub fn shelf(&mut self, a: T) -> &mut Self

Set linear shelf gain

Used only for peaking, highshelf, lowshelf filters.

Arguments

a: Linear shelf gain. Defaults to 1.0.

pub fn shelf_db(&mut self, a_db: T) -> &mut Self

Set shelf gain in dB

Used only for peaking, highshelf, lowshelf filters.

Arguments

a_db: Linear shelf gain. Defaults to 0.0.

pub fn inverse_q(&mut self, qi: T) -> &mut Self

Set inverse Q parameter of the filter

The inverse “steepness”/“narrowness” of the filter transition. Defaults sqrt(2) which is as steep as possible without ripple.

Arguments

qi: Inverse Q parameter.

pub fn q(&mut self, q: T) -> &mut Self

Set Q parameter of the filter

The “steepness”/“narrowness” of the filter transition. Defaults 1/sqrt(2) which is as steep as possible without ripple.

This affects the same parameter as bandwidth() and shelf_slope(). Use only one of them.

Arguments

q: Q parameter.

pub fn bandwidth(&mut self, bw: T) -> &mut Self

Set the relative bandwidth

This affects the same parameter as inverse_q() and shelf_slope(). Use only one of them.

Arguments

bw: Bandwidth in octaves

pub fn shelf_slope(&mut self, s: T) -> &mut Self

Set the shelf slope.

This affects the same parameter as inverse_q() and bandwidth(). Use only one of them.

Arguments

s: Shelf slope. A slope of 1.0 is maximally steep without overshoot.

pub fn set_shape(&mut self, s: Shape<T>) -> &mut Self

Set the shape

pub fn lowpass(&self) -> [[T; 3]; 2]

Low pass filter

Builds second order biquad low pass filter coefficients.

use dsp_fixedpoint::Q32;
use dsp_process::SplitInplace;
use idsp::iir::*;
let iir: Biquad<Q32<30>> = coefficients::Filter::default()
    .critical_frequency(0.1)
    .gain(1000.0)
    .lowpass()
    .into();
let mut xy = vec![3, -4, 5, 7, -3, 2];
iir.inplace(&mut DirectForm1::default(), &mut xy);
assert_eq!(xy, [5, 3, 9, 25, 42, 49]);

pub fn highpass(&self) -> [[T; 3]; 2]

High pass filter

Builds second order biquad high pass filter coefficients.

use dsp_fixedpoint::Q32;
use dsp_process::SplitInplace;
use idsp::iir::*;
let iir: Biquad<Q32<30>> = coefficients::Filter::default()
    .critical_frequency(0.1)
    .gain(1000.0)
    .highpass()
    .into();
let mut xy = vec![3, -4, 5, 7, -3, 2];
iir.inplace(&mut DirectForm1::default(), &mut xy);
assert_eq!(xy, [5, -9, 11, 12, -1, 17]);

pub fn bandpass(&self) -> [[T; 3]; 2]

Band pass

use idsp::iir::*;
let ba = coefficients::Filter::default()
    .frequency(1000.0, 48e3)
    .q(5.0)
    .gain_db(3.0)
    .bandpass();
println!("{ba:?}");

pub fn notch(&self) -> [[T; 3]; 2]

A notch filter

Has zero gain at the critical frequency.

pub fn allpass(&self) -> [[T; 3]; 2]

An allpass filter

Has constant gain at all frequencies but a variable phase shift.

pub fn peaking(&self) -> [[T; 3]; 2]

A peaking/dip filter

Has gain*shelf_gain at critical frequency and gain elsewhere.

pub fn lowshelf(&self) -> [[T; 3]; 2]

Low shelf

Approaches gain*shelf_gain below critical frequency and gain above.

use idsp::iir::*;
let ba = coefficients::Filter::default()
    .frequency(1000.0, 48e3)
    .shelf_slope(2.0)
    .shelf_db(20.0)
    .lowshelf();
println!("{ba:?}");

pub fn highshelf(&self) -> [[T; 3]; 2]

Low shelf

Approaches gain*shelf_gain above critical frequency and gain below.

pub fn iho(&self) -> [[T; 3]; 2]

I/HO

Notch, integrating below, flat shelf_gain above