Struct idsp::iir::IIR

pub struct IIR<T> {
    pub ba: Vec5<T>,
    pub y_offset: T,
    pub y_min: T,
    pub y_max: T,
}

IIR configuration.

Contains the coeeficients ba, the output offset y_offset, and the output limits y_min and y_max. Data is represented in variable precision floating-point. The dataformat is the same for all internal signals, input and output.

This implementation achieves several important properties:

Its transfer function is universal in the sense that any biquadratic transfer function can be implemented (high-passes, gain limits, second order integrators with inherent anti-windup, notches etc) without code changes preserving all features.

It inherits a universal implementation of “integrator anti-windup”, also and especially in the presence of set-point changes and in the presence of proportional or derivative gain without any back-off that would reduce steady-state output range.

It has universal derivative-kick (undesired, unlimited, and un-physical amplification of set-point changes by the derivative term) avoidance.

An offset at the input of an IIR filter (a.k.a. “set-point”) is equivalent to an offset at the output. They are related by the overall (DC feed-forward) gain of the filter.

It stores only previous outputs and inputs. These have direct and invariant interpretation (independent of gains and offsets). Therefore it can trivially implement bump-less transfer.

Cascading multiple IIR filters allows stable and robust implementation of transfer functions beyond bequadratic terms.

Miniconf

{"y_offset": y_offset, "y_min": y_min, "y_max": y_max, "ba": [b0, b1, b2, a1, a2]}

y0 is the output offset code

ym is the lower saturation limit

yM is the upper saturation limit

IIR filter tap gains (ba) are an array [b0, b1, b2, a1, a2] such that the new output is computed as y0 = a1*y1 + a2*y2 + b0*x0 + b1*x1 + b2*x2. The IIR coefficients can be mapped to other transfer function representations, for example as described in https://arxiv.org/abs/1508.06319

Fields
ba: Vec5<T>y_offset: Ty_min: Ty_max: T
Implementations

impl<T: Float + Default + Sum<T>> IIR<T>

pub fn new(gain: T, y_min: T, y_max: T) -> Self

pub fn set_pi(&mut self, kp: T, ki: T, g: T) -> Result<(), &str>

Configures IIR filter coefficients for proportional-integral behavior with gain limit.

Arguments

kp - Proportional gain. Also defines gain sign.

ki - Integral gain at Nyquist. Sign taken from kp.

g - Gain limit.

pub fn get_k(&self) -> T

Compute the overall (DC feed-forward) gain.

pub fn get_x_offset(&self) -> Result<T, &str>

pub fn set_x_offset(&mut self, xo: T)

Convert input (x) offset to equivalent output (y) offset and apply.

Arguments

xo: Input (x) offset.

pub fn update(&self, xy: &mut Vec5<T>, x0: T, hold: bool) -> T

Feed a new input value into the filter, update the filter state, and return the new output. Only the state xy is modified.

Arguments

xy - Current filter state.

x0 - New input.