idsp

Trait Build

pub trait Build<C> {
    type Context;

    // Required method
    fn build(&self, ctx: &Self::Context) -> C;
}

Build a value using a context

This is similar to the Into trait but allows lossy (rounding, clamping, quantization) and non-value-preserving conversions. The only semantic constraint is that the conversion is obvious and unambiguous.

Required Associated Types

type Context

The context of the convesion.

E.g. units.

Required Methods

fn build(&self, ctx: &Self::Context) -> C

Perform the conversion