idsp

Trait Filter

pub trait Filter {
    type Config;

    // Required methods
    fn update(&mut self, x: i32, k: &Self::Config) -> i32;
    fn get(&self) -> i32;
    fn set(&mut self, x: i32);
}

Single inpout single output i32 filter

Required Associated Types

type Config

Filter configuration type.

While the filter struct owns the state, the configuration is decoupled to allow sharing.

Required Methods

fn update(&mut self, x: i32, k: &Self::Config) -> i32

Update the filter with a new sample.

Args

x: Input data.

k: Filter configuration.

Return

Filtered output y.

fn get(&self) -> i32

Return the current filter output

fn set(&mut self, x: i32)

Update the filter so that it outputs the provided value. This does not completely define the state of the filter.